Add @angular/forms to package.json
In RC3, the forms module is now now a separate package. It stays so with RC4.
Forms are quite essential for Angular2 though, and were for long part of the main packages bundle, so, I thought this doesn't need to be its own `angularPackages()` thing and just added it.

The forms version 0.1.1 matches RC3 and was not included.
The forms version 0.2.0 matches RC4 and this is what's added here.